Update pageviews top and by-country response def
The top-endpoint used to send a string as a response (a json formatted
string, but a string nonetheless). When restbase-cassandra got a JSON
type, we updated the actual response but forgot to update the definition.
Here it is updated, in addition to the by-country one, that follows the
same logic.

Bug: T184541
Change-Id: I16183b2040cf89419f70600ce0609122da6846b4