Add a config param for druid datasources
To improve resiliency of deploy new data in Druid for AQS to serve,
we will keep multiple versions of each datasource as we update
them, facilitating swaping between them and rollbacking.
The configuration parameter for datasources allows for manual
swapping through a puppet merge and an AQS restart (change
of config file managed by puppet).

Bug: T193387
Change-Id: I2c2d1127452b4f5f24a9d998224955802b0defb9