Update new_pages endpoint druid query
Currently the druid query for new_pages statistis computes
(new-pages events + restore-page events - delete events)
However as of now we do not have delete events in the dataset,
and restore-events are revision oriented (meaning they occur when
revisions are restored onto a page, which in case doesnt exist,
gets created with a reate event).
This patch makes the druid query only use create-events.

Bug: T200272
Change-Id: I2706c97b6d943d4bbf65c1e04cc556c9d86c28ee