Refactor test runner and fix tests stalling
In the current state of the tests we're opening a new test runner for each
test group. This change refactors the service runner client for the tests,
declares the tests to be run explicitly and eliminates some logic that's out
of use.

This also means that the project is ready to be CI'd by Jenkins.

Bug: T209711
Change-Id: Id9fedbdac713f8589673e69d583b13e28a08b9d3