Url encode file name before querying the data store
The mediarequest/mediacounts dataset stores file names in a "particular"
way, where path slashes are not url encoded but the names themselves are,
example:

/wiki/commons/5/53/Pr%C3%A4sidentschaftswahl_in_den_Vereinigten_Staaten.ogv

Paths in the GET requests are indeed url encoded, but they are delivered
to AQS as decoded strings. For the internal storage request to find them,
therefore, we need to re-urlencode the file name so that it matches how it
is stored in Cassandra.

Bug: T244373
Change-Id: Ie467a4cdd62ec30f7cdad8f24e0831b4e1bcbc95