AQS: add configuration for timeout to Druid requests
Currently the timeout for HTTP requests from AQS into the Druid LVS
is too high, which means connections end up piling up and displaying
some degradation at the Wikistats UI level.

This change adds parameters to the Hyperswitch request config to add
both an explicit timeout of 10 seconds to Druid requests, as well as a
(commented out for now) maximum retries value of 3.

Bug: T268809
Change-Id: I318fcb966fd19ba552a15a385069c955606d307d