Add makefile and dockerfile for local tests
Currently our documentation recommends running the aqs tests locally
when developing. Given that there's a high chance of developers
running tests locally on OS X, let's use a docker container to run the
tests within something close to the WMF enviornment.

In future we could use a docker-compose setup to run a cassandra
instance locally also. Default to buster - currently tests pass on
buster and we'll be running it soon anyway.

Change-Id: Ic746540d202e489ec8ab65669eb8ebff84118c66