Fix build by bumping versions
* Service runner defaults the deployment target to jessie, which errors
out with 404s during apt install commands run from in docker.
Explicitly passing debian:bullseye allows us to keep building AQS for
hopefully a little more time
* The current service-runner dependency tries to install nodejs-legacy,
which is not available since debian:jessie, which in turn is not
available at all.  Bumping service-runner to 3.1.0 seems to work and
gets rid of that dependency

Change-Id: I6af0e65c288124fad0e104522b13a2d73a600fa3